+++ src/helpers/configuration_helpers.sh
@@ -20,7 +20,7 @@ function add_perennial_branch {
     echo_inline_error "'$branch_name' is already set as the main branch"
     exit_with_error
   else
-    local new_branches=$(insert_string "$PERENNIAL_BRANCH_NAMES" ',' "$branch_name")
+    local new_branches=$(insert_string "$PERENNIAL_BRANCH_NAMES" ' ' "$branch_name")
     store_configuration perennial-branch-names "$new_branches"
   fi
 }
@@ -43,22 +43,18 @@ function add_or_remove_perennial_branches {
 }
 
 
-function echo_perennial_branch_usage {
-  echo_inline_usage 'git town perennial-branches (--add | --remove) <branch_name>'
+# Returns whether or not the perennial branches are configured
+function are_perennial_branches_configured {
+  if get_configuration 'perennial-branch-names'; then
+    echo true
+  else
+    echo false
+  fi
 }
 
 
-# Ensure that perennial branches don't contain main branch
-function ensure_valid_perennial_branches {
-  local branches=$1
-
-  split_string "$branches" ',' | while read branch; do
-    if [[ "$branch" == "$MAIN_BRANCH_NAME" ]]; then
-      echo_error_header
-      echo_error "'$branch' is already set as the main branch"
-      exit_with_error
-    fi
-  done
+function echo_perennial_branch_usage {
+  echo_inline_usage 'git town perennial-branches (--add | --remove) <branch_name>'
 }
 
 
@@ -69,9 +65,9 @@ function get_configuration {
 }
 
 
-# Returns whether or not Git Town is configured
-function is_git_town_configured {
-  if [ -n "$MAIN_BRANCH_NAME" ] && get_configuration 'perennial-branch-names'; then
+# Returns whether or not the main branch is configured
+function is_main_branch_configured {
+  if [ -n "$MAIN_BRANCH_NAME" ]; then
     echo true
   else
     echo false
@@ -99,62 +95,19 @@ function remove_perennial_branch {
     echo_inline_error "'$branch_name' is not a perennial branch"
     exit_with_error
   else
-    local new_branches=$(remove_string "$PERENNIAL_BRANCH_NAMES" ',' "$branch_name")
+    local new_branches=$(remove_string "$PERENNIAL_BRANCH_NAMES" ' ' "$branch_name")
     store_configuration perennial-branch-names "$new_branches"
   fi
 }
 
 
-# Begin the git town setup wizard
-function setup_configuration {
-  setup_configuration_main_branch
-  echo
-  setup_configuration_perennial_branches
-  echo "Done with configuration:"
-  show_config | indent
-}
-
-
-# Ask and store main-branch-name
-function setup_configuration_main_branch {
-  echo "Please specify the main dev branch (typically 'master' or 'development'):"
-  read main_branch_input
-  if [[ -z "$main_branch_input" ]]; then
-    echo_error_header
-    echo_error "You have not provided the name for the main branch."
-    echo_error "Aborting Git Town configuration."
-    exit_with_error newline
-  fi
-
-  ensure_has_branch "$main_branch_input" || exit_with_error
-  store_configuration main-branch-name "$main_branch_input"
-}
-
-
-# Ask and store perennial-branch-names
-function setup_configuration_perennial_branches {
-  echo "Git Town supports perennial branches like 'release' or 'production'."
-  echo "These branches cannot be shipped and will not merge '$MAIN_BRANCH_NAME' when syncing."
-  echo "Please enter your perennial branches as a comma separated list or a blank line to skip."
-  echo "Example: 'qa, production'"
-  read perennial_input
-
-  if [[ -n "$perennial_input" ]]; then
-    ensure_has_branches "$perennial_input" || exit_with_error
-    ensure_valid_perennial_branches "$perennial_input" || exit_with_error
-  fi
-
-  store_configuration perennial-branch-names "$perennial_input"
-}
-
-
-# Perform `git town config` operation ("reset", "setup", "show")
+# Perform `git town config` operation
 function run_config_operation {
   local operation=$1
 
   if [ -n "$operation" ]; then
     if [ "$operation" == "--setup" ]; then
-      setup_configuration
+      ensure_knows_configuration
     elif [ "$operation" == "--reset" ]; then
       remove_all_configuration
     else
@@ -172,7 +125,7 @@ function show_config {
   echo_inline_bold "Perennial branches:"
   if [ -n "$PERENNIAL_BRANCH_NAMES" ]; then
     echo
-    split_string "$PERENNIAL_BRANCH_NAMES" ","
+    split_string "$PERENNIAL_BRANCH_NAMES" " "
   else
     echo ' [none]'
   fi
@@ -190,7 +143,7 @@ function show_main_branch {
 
 function show_perennial_branches {
   if [ -n "$PERENNIAL_BRANCH_NAMES" ]; then
-    split_string "$PERENNIAL_BRANCH_NAMES" ","
+    split_string "$PERENNIAL_BRANCH_NAMES" " "
   fi
 }